requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
bcrypt==4.0.1
PyJWT
pydantic
python-dotenv
email-validator
